+++ src/consensus/consensus_types.h
@@ -10,15 +10,6 @@
 
 #include <stdint.h>
 
-namespace ccf::consensus
-{
-  DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(Configuration);
-  DECLARE_JSON_REQUIRED_FIELDS(Configuration);
-  DECLARE_JSON_OPTIONAL_FIELDS(
-    Configuration, message_timeout, election_timeout, max_uncommitted_tx_count);
-
-}
-
 namespace consensus
 {
 #pragma pack(push, 1)