+++ src/node/rpc/node_interface.h
@@ -4,10 +4,10 @@
 
 #include "ccf/crypto/pem.h"
 #include "ccf/ds/quote_info.h"
+#include "ccf/node/configuration.h"
 #include "ccf/node_startup_state.h"
 #include "ccf/service/node_info_network.h"
 #include "ccf/service/tables/code_id.h"
-#include "common/configuration.h"
 #include "http/http_builder.h"
 #include "http/http_parser.h"
 #include "kv/store.h"
@@ -58,7 +58,8 @@ namespace ccf
     virtual RecoveryDecisionProtocolSubsystem&
     get_recovery_decision_protocol() = 0;
     virtual void shuffle_sealed_shares(ccf::kv::Tx& tx) = 0;
-    [[nodiscard]] virtual const ccf::StartupConfig& get_node_config() const = 0;
+    [[nodiscard]] virtual const ccf::CCFConfig& get_node_config() const = 0;
+    [[nodiscard]] virtual const nlohmann::json& get_node_data() const = 0;
     virtual ccf::crypto::Pem get_network_cert() = 0;
     virtual void stop_notice() = 0;
     virtual bool has_received_stop_notice() = 0;